cloud_functions/consumer/requirements.txt
vaderSentiment==3.3.2
google-cloud-spanner==3.3.0
